client

+++ nets_lib/receivenbytes.c
@@ -4,7 +4,7 @@
 
 #include <sys/socket.h>
 
-long receiveNBytes(int socket, size_t n, void *bytes) {
+ssize_t receiveNBytes(int socket, size_t n, void *bytes) {
     size_t received = 0;
     while (received < n && received >= 0) {
         ssize_t recvd = recv(socket, &bytes[received], n - received, 0);
